Make sure the Python code supports Torch v1.11

+++ python/metatensor-torch/tests/operations/one_hot.py
@@ -11,7 +11,7 @@ def test_one_hot():
         values=torch.tensor([[0, 6], [1, 1], [4, 6]]),
     )
     possible_labels = metatensor.torch.Labels("species", torch.tensor([[1], [6]]))
-    correct_encoding = torch.tensor([[0, 1], [1, 0], [0, 1]], dtype=torch.int32)
+    correct_encoding = torch.tensor([[0, 1], [1, 0], [0, 1]])
     one_hot_encoding = metatensor.torch.one_hot(original_labels, possible_labels)
 
     assert isinstance(one_hot_encoding, torch.Tensor)
